+++ src-docs/src/routes.js
@@ -204,6 +204,9 @@ import { ToastExample }
 import { ToolTipExample }
   from './views/tool_tip/tool_tip_example';
 
+import { ToggleExample }
+  from './views/toggle/toggle_example';
+
 import { Changelog }
   from './views/package/changelog';
 
@@ -344,6 +347,7 @@ const navigation = [{
     IsColorDarkExample,
     OutsideClickDetectorExample,
     PortalExample,
+    ToggleExample,
     UtilityClassesExample,
   ].map(example => createExample(example)),
 }, {