+++ src-docs/src/views/button/button_example.js
@@ -11,6 +11,9 @@ import {
   EuiButtonEmpty,
   EuiButtonIcon,
   EuiCode,
+  EuiButtonGroup,
+  EuiButtonToggle,
+  EuiLink,
 } from '../../../../src/components';
 
 import Button from './button';
@@ -44,6 +47,14 @@ const buttonAsLinkHtml = renderToHtml(ButtonAsLink);
 import ButtonLoading from './button_loading';
 const buttonLoadingSource = require('!!raw-loader!./button_loading');
 const buttonLoadingHtml = renderToHtml(ButtonLoading);
+
+import ButtonToggle from './button_toggle';
+const buttonToggleSource = require('!!raw-loader!./button_toggle');
+const buttonToggleHtml = renderToHtml(ButtonToggle);
+
+import ButtonGroup from './button_group';
+const buttonGroupSource = require('!!raw-loader!./button_group');
+const buttonGroupHtml = renderToHtml(ButtonGroup);
 
 export const ButtonExample = {
   title: 'Button',
@@ -159,7 +170,60 @@ export const ButtonExample = {
     props: { EuiButtonIcon },
     demo: <ButtonIcon />,
   }, {
-    title: 'Ghost buttons for deep color backgrounds',
+    title: 'Toggle buttons',
+    source: [{
+      type: GuideSectionTypes.JS,
+      code: buttonToggleSource,
+    }, {
+      type: GuideSectionTypes.HTML,
+      code: buttonToggleHtml,
+    }],
+    text: (
+      <div>
+        <p>
+          This is a specialized component that combines <EuiCode>EuiButton</EuiCode> and <EuiCode>EuiToggle</EuiCode> to
+          create a button with an on/off state. You can pass all the same parameters to it as you
+          can to <EuiCode>EuiButton</EuiCode>. The main difference is that, it does not accept any children,
+          but a <EuiCode>label</EuiCode> prop instead. This is for the handling of accessiblity with
+          the <EuiCode>EuiToggle</EuiCode>.
+        </p>
+        <p>
+          The <EuiCode>EuiButtonToggle</EuiCode> does not have any inherit visual state differences.
+          These you must apply in your implementation.
+        </p>
+      </div>
+    ),
+    demo: <ButtonToggle />,
+    props: { EuiButtonToggle },
+  }, {
+    title: 'Groups',
+    source: [{
+      type: GuideSectionTypes.JS,
+      code: buttonGroupSource,
+    }, {
+      type: GuideSectionTypes.HTML,
+      code: buttonGroupHtml,
+    }],
+    text: (
+      <div>
+        <p>
+          Button groups are handled similarly to the way checkbox and radio groups are
+          handled but made to look like buttons. They group multiple <EuiCode>EuiButtonToggle</EuiCode>s
+          and utilize the <EuiCode>type=&quot;single&quot;</EuiCode> or <EuiCode>&quot;multi&quot;</EuiCode> prop
+          to determine whether multiple or only single selections are allowed per group.
+        </p>
+        <p>
+          Stylistically, all button groups are the size of small buttons, do not stretch to fill the container
+          and typically should only be <EuiCode>color=&quot;text&quot;</EuiCode> (default)
+          or <EuiCode>&quot;primary&quot;</EuiCode>.
+          If your just displaying a group of icons, add the prop <EuiCode>isIconOnly</EuiCode>.
+        </p>
+      </div>
+    ),
+    demo: <ButtonGroup />,
+    props: { EuiButtonGroup },
+  }, {
+    title: 'Ghost',
     source: [{
       type: GuideSectionTypes.JS,
       code: buttonGhostSource,
@@ -169,11 +233,10 @@ export const ButtonExample = {
     }],
     text: (
       <p>
-        You can also pass <EuiCode>type=&apos;ghost&apos;</EuiCode> to any of the button
-        styles on this page. These should be used extremely rarely, and are
+        For buttons on dark color backgrounds, you can pass <EuiCode>color=&apos;ghost&apos;</EuiCode> to
+        any of the button styles on this page. These should be used extremely rarely, and are
         only for placing buttons on top of dark or image-based backgrounds.
-        A good example of their use is in
-        the <EuiCode>EuiBottomBar</EuiCode> component
+        A good example of their use is in the <EuiLink href="/#/layout/bottom-bar">EuiBottomBar</EuiLink> component.
       </p>
     ),
     demo: <ButtonGhost />,