+++ src/components/button/button_toggle/index.js
@@ -1,0 +1,3 @@
+export {
+  EuiButtonToggle,
+} from './button_toggle';