+++ src/components/toggle/toggle.test.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { render } from 'enzyme';
+import { requiredProps } from '../../test';
+
+import { EuiToggle } from './toggle';
+
+describe('EuiToggle', () => {
+  test('is rendered', () => {
+    const component = render(
+      <EuiToggle label="Is toggle on?" {...requiredProps} />
+    );
+
+    expect(component)
+      .toMatchSnapshot();
+  });
+});